chore: rename project
