+++ .github/release-please/.manifest.json
@@ -1,4 +1,4 @@
 {
-    ".": "0.2.3",
-    "fakerator-core": "0.2.0"
+    ".": "0.2.2",
+    "fauxgen-core": "0.2.0"
 }
