+++ .github/workflows/lint-and-test.yaml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: ./fakerator-core
+        working-directory: ./fauxgen-core
     steps:
       - name: Checkout sources
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
@@ -22,7 +22,7 @@ jobs:
       - uses: Swatinem/rust-cache@9d47c6ad4b02e050fd481d890b2ea34778fd09d6 # v2.7.8
         with:
           workspaces: |
-            fakerator-core -> fakerator-core-target
+            fauxgen-core -> fauxgen-core-target
 
       - name: Run check
         run: cargo check --all-targets --all-features
@@ -35,7 +35,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: ./fakerator-core
+        working-directory: ./fauxgen-core
     steps:
       - name: Checkout sources
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
@@ -43,7 +43,7 @@ jobs:
       - uses: Swatinem/rust-cache@9d47c6ad4b02e050fd481d890b2ea34778fd09d6 # v2.7.8
         with:
           workspaces: |
-            fakerator-core -> fakerator-core-target
+            fauxgen-core -> fauxgen-core-target
 
       - name: Run fmt
         run: cargo fmt --all --check
