+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["fakerator"]
+  "cSpell.words": ["fauxgen"]
 }
