+++ README.md
@@ -1,23 +1,23 @@
-# fakerator
+# fauxgen
 
 A factory method generator that streamlines test data creation by automating field value generation.
 
 ## Overview
 
-Testing classes with numerous fields often requires extensive boilerplate code. `fakerator` addresses this challenge by automatically generating factory methods from class definitions, significantly reducing test code complexity.
+Testing classes with numerous fields often requires extensive boilerplate code. `fauxgen` addresses this challenge by automatically generating factory methods from class definitions, significantly reducing test code complexity.
 
 Currently supports `pandera.DataFrameModel` with plans for future expansions.
 
 ## Quick Start
 
 ```bash
-uv add fakerator
-uv run fakerator gen --module-dir <your_module>
+uv add fauxgen
+uv run fauxgen gen --module-dir <your_module>
 ```
 
 ## Usage Example
 
-### Traditional Approach (Without fakerator)
+### Traditional Approach (Without fauxgen)
 
 Testing DataFrame models traditionally requires explicit specification of all fields, even when testing a single validation:
 
@@ -46,12 +46,12 @@ def test_user_registration():
     # Test age validation...
 ```
 
-### Simplified Testing (With fakerator)
+### Simplified Testing (With fauxgen)
 
-fakerator generates factory methods that enable focused testing by automatically handling irrelevant fields:
+fauxgen generates factory methods that enable focused testing by automatically handling irrelevant fields:
 
 ```python
-from .factories import user_schema_record  # Generated by fakerator
+from .factories import user_schema_record  # Generated by fauxgen
 
 def test_user_registration():
     # Focus solely on the field under test
@@ -86,7 +86,7 @@ class TestSchema(pa.DataFrameModel):
 ### Output: Generated Factory Method
 
 ```python
-import fakerator as f
+import fauxgen as f
 
 class TestSchemaRecord(TypedDict):
     int_col: int
