+++ examples/pyproject.toml
@@ -1,16 +1,16 @@
 [project]
 dependencies = [
   "pandera[mypy]>=0.23.1",
-  "fakerator",
+  "fauxgen",
 ]
-description = "examples projects for fakerator"
+description = "examples projects for fauxgen"
 name = "examples"
 readme = "README.md"
 requires-python = ">=3.12"
 version = "0.1.0"
 
 [tool.uv.sources]
-fakerator = {path = "../"}
+fauxgen = {path = "../", editable = true}
 
 [dependency-groups]
 dev = [
