+++ examples/tests/test_service.py
@@ -4,7 +4,7 @@ from pandera.typing import DataFrame
 
 from src.schema.user import UserSchema
 from src.service import exclude_under_20_users
-from src.testing.fakerator.schema.user import user_schema_record
+from src.testing.fauxgen.schema.user import user_schema_record
 
 
 @pytest.mark.parametrize(
