+++ fauxgen-core/CHANGELOG.md
@@ -1,0 +1,20 @@
+# Changelog
+
+## [0.2.0](https://github.com/hiro-o918/fauxgen/compare/0.1.0...0.2.0) (2025-03-29)
+
+### Features
+
+- add cli for fauxgen ([#3](https://github.com/hiro-o918/fauxgen/issues/3)) ([18ab5f9](https://github.com/hiro-o918/fauxgen/commit/18ab5f944d844501817eef435d110e055568063e))
+- add examples and readme ([#4](https://github.com/hiro-o918/fauxgen/issues/4)) ([8c79f5c](https://github.com/hiro-o918/fauxgen/commit/8c79f5ce34b6a4baf465aa1bd30d14f6ae0a9113))
+- add factory for pandera dataframe ([fb0557f](https://github.com/hiro-o918/fauxgen/commit/fb0557f18767d85b01394d1add8c9791a76b23ce))
+- add fauxgen-core ([dbb88b1](https://github.com/hiro-o918/fauxgen/commit/dbb88b195dd0272ece7d686574c3010adee52dfa))
+- add generator for a file ([a283ca8](https://github.com/hiro-o918/fauxgen/commit/a283ca8dc365218923631943de0d1eb0b846b83e))
+- initialize cargo project ([b69f11f](https://github.com/hiro-o918/fauxgen/commit/b69f11f5dccdf4845d9cfefc1a5ee941bbe4c054))
+- initialize cargo project ([6777c68](https://github.com/hiro-o918/fauxgen/commit/6777c68a3248d19c9fad15ec391f06acda0bed5a))
+- write files to factory dir ([8e56be8](https://github.com/hiro-o918/fauxgen/commit/8e56be868c89b3d95be8c8fc782e606b8e53db0d))
+
+### Miscellaneous
+
+- add some deps ([eeb41ab](https://github.com/hiro-o918/fauxgen/commit/eeb41abb2f97af3b51aff7cb3026f7bfd62b4fbc))
+- fix lint ([94c2a09](https://github.com/hiro-o918/fauxgen/commit/94c2a097d8a52f76c9ff2ce17205838221b7da8c))
+- move bin ([21f9d9e](https://github.com/hiro-o918/fauxgen/commit/21f9d9e48d6ff3046ff3b0e730e1d510ff55fa70))
