+++ fauxgen-core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 edition = "2021"
-name = "fakerator-core"
+name = "fauxgen-core"
 version = "0.2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -14,7 +14,7 @@ regex = "1.11.1"
 rustpython-parser = "0.4.0"
 
 [[bin]]
-name = "fakerator"
+name = "fauxgen"
 path = "src/main.rs"
 
 [dev-dependencies]
