+++ fauxgen-core/resources/.python-version
