+++ fauxgen-core/resources/generator/render_factory_code_from_file/pandera_field_parameters_expected.py
@@ -1,7 +1,7 @@
 import datetime
 from typing import TypedDict
 
-import fakerator as f
+import fauxgen as f
 
 
 class TestSchemaRecord(TypedDict):
