+++ fauxgen-core/resources/generator/write_factory_codes/expected/bar/__init__.py
@@ -1,7 +1,7 @@
 import datetime
 from typing import TypedDict
 
-import fakerator as f
+import fauxgen as f
 
 
 class BarSchemaRecord(TypedDict):
