+++ fauxgen-core/resources/generator/write_factory_codes/input/bar/__init__.py
