+++ fauxgen-core/resources/generator/write_factory_codes/input/foo.py
