+++ fauxgen-core/resources/pyproject.toml
