+++ fauxgen-core/src/lib.rs
