+++ fauxgen-core/src/strings.rs
