+++ fauxgen-core/templates/pandera_factory.py.tpl
