+++ pyproject.toml
@@ -1,38 +1,34 @@
 [project]
-name = "fakerator"
-version = "0.2.3"
-description = "Add your description here"
-readme = "README.md"
 authors = [
-    { name = "Hironori Yamamoto", email = "mr.nikoru918@gmail.com" }
+  {name = "Hironori Yamamoto", email = "mr.nikoru918@gmail.com"},
 ]
+description = "Add your description here"
+name = "fauxgen"
+readme = "README.md"
 requires-python = ">=3.10"
-dependencies = [
-    "click>=8.1.8",
-]
+version = "0.2.2"
 
 [tool.uv]
 package = true
 
 [tool.maturin]
-manifest-path = "fakerator-core/Cargo.toml"
 bindings = "bin"
+exclude = [
+  "fauxgen_core/resources/**/*",
+]
+manifest-path = "fauxgen-core/Cargo.toml"
 python-source = "python"
 strip = true
-exclude = [
-    "fakerator_core/resources/**/*",
-]
 
 [build-system]
-requires = ["maturin>=1.0,<2.0"]
 build-backend = "maturin"
+requires = ["maturin>=1.0,<2.0"]
 
 [dependency-groups]
 dev = [
-    "mypy>=1.15.0",
-    "ruff>=0.11.2",
+  "mypy>=1.15.0",
+  "ruff>=0.11.2",
 ]
-
 
 [tool.ruff]
 # Exclude a variety of commonly ignored directories.
@@ -63,8 +59,8 @@ exclude = [
   "node_modules",
   "site-packages",
   "venv",
-  "fakerator-core/resources/",
-  "fakerator-core/templates/",
+  "fauxgen-core/resources/",
+  "fauxgen-core/templates/",
 ]
 
 # Same as Black.
@@ -116,7 +112,7 @@ docstring-code-format = false
 docstring-code-line-length = "dynamic"
 
 [tool.mypy]
-exclude =  [".venv/", "fakerator-core/resources/"]
+exclude = [".venv/", "fauxgen-core/resources/"]
 
 check_untyped_defs = true
 disallow_any_decorated = false
